Add files via upload
In this project, using the key functions of SQL I was able to perform data exploration on the Covid19 data set. The data set includes the data from 2020-2021, the two peak years of the COVID-19 pandemic. 
Many of my concepts of SQL such as: aggregate functions, joins, where clause, having clause, group by, order by and Common table expressions were applied while exploration of this dataset. The data was able to giveaway many talking points which can be looked upon to reach a conclusion that how the vaccinations process was carried out with respect to each country. Some numbers depicted that stringency index of a country may or may not have had an impact when trying to control the pandemic affected areas. It improved my ability to identify and to explore data in such a way that it is useful in the overall context.
The data set is openly available at the following link:  https://ourworldindata.org/covid-deaths